+++ CHANGES.rst
@@ -39,13 +39,16 @@ New Features
 Fixes
 -----
 
-* newSpecFile did not update scan_id in some cases.
+* Application new_bluesky_environment would crash when requests package not found.
+* Function newSpecFile did not update scan_id in some cases.
 
 Maintenance
 ------------
 
 * Add environment file for 2024_2 cycle.
+* Testing with Python 3.9, 3.10, 3.11.
 * Make the home page less complex.
+* Require Py3.9 or higher to run new_bluesky_environment application.
 
 1.0.2
 ******